backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
tortoise-orm
pydantic[email]
PyJWT
python-dotenv
aiosqlite 
pydantic-settings